requirements.txt
Flask>=3.0.0
Flask-Cors>=4.0.0
Flask-SQLAlchemy>=3.1.0
PyJWT>=2.8.0
python-dotenv>=1.0.0
gunicorn>=22.0.0
psycopg2-binary>=2.9.9
